[ISSUE-1400] Share email invitation sending logic between test classes
